requirements.txt
pybit==5.6.0
websocket-client==1.6.4
asyncio==3.4.3
aiohttp==3.9.1
aiohttp-cors==0.7.0
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
numpy==1.24.3
pandas==2.0.3
schedule==1.2.0
psutil==5.9.5
redis==4.6.0 